JOBS CRUD finished
Workers can now apply for and withdraw from jobs.
Clients have the ability to designate a final applicant for a job and mark it as completed.
Rules have been established for both Clients and Workers, including restrictions on editing or deleting ongoing work. These rules are enforced using the "auth" middleware, which considers the user's type during authorization checks.